Warning fixed
svn path=/trunk/vorbis-tools/; revision=14564

+++ oggdec/oggdec.c
@@ -341,7 +341,7 @@ static int decode_file(FILE *in, FILE *out, char *infile, char *outfile)
         if(!quiet && seekable) {
             done += ret/size;
             if((double)done/(double)length * 200. > (double)percent) {
-                percent = (double)done/(double)length *200;
+                percent = (int)((double)done/(double)length *200);
                 fprintf(stderr, "\r\t[%5.1f%%]", (double)percent/2.);
             }
         }
